Add auto-postings to the accounts document
Auto postings are posting that only have an account id. The amount and
currency are inferred from the other postings.

As part of this change we change the add transaction method of the
accounts doc to take a vec of postings rather than a transaction. This
means all the validity checks for the postings/transactions are done in
one place rather than being spread between add postings to the tx and
adding the tx to the accounts doc. This change was made because with
auto-postings there are more checks and the checks are more complicated
so it's simpler/cleaner to do them in one pass.